Move physical properties into separate mod
