+++ rbx_types/src/lib.rs
@@ -7,6 +7,7 @@ mod basic_types;
 mod brick_color;
 mod faces;
 mod lister;
+mod physical_properties;
 mod shared_string;
 mod variant;
 
@@ -14,5 +15,6 @@ pub use axes::*;
 pub use basic_types::*;
 pub use brick_color::*;
 pub use faces::*;
+pub use physical_properties::*;
 pub use shared_string::*;
 pub use variant::*;
